[fix] 매일지출내역 로그인 정보 조회할 수 있게 코드 수정

+++ src/main/java/com/zerobase/accountbook/service/dailypaymetns/DailyPaymentsService.java
@@ -54,15 +54,16 @@ public class DailyPaymentsService {
 
     @CachePut(value = "dailyPayments", key = "#request.dailyPaymentsId")
     public ModifyDailyPaymentsResponseDto modifyDailyPayments(
+            String memberEmail,
             ModifyDailyPaymentsRequestDto request
     ) {
 
-        Member member = validateMember(request.getMemberEmail());
+        Member member = validateMember(memberEmail);
         DailyPayments dailyPayments =
                 validateDailyPayments(request.getDailyPaymentsId());
 
         // 지출내역의 주인과 수정하려는 사용자가 다를 경우
-        requestMemberAndDailyPaymentsOwnerMismatch(member, dailyPayments);
+        checkDailyPaymentsOwner(member, dailyPayments);
 
         dailyPayments.setPaidAmount(request.getPaidAmount());
         dailyPayments.setPaidWhere(request.getPaidWhere());
@@ -77,28 +78,40 @@ public class DailyPaymentsService {
     }
 
     @CacheEvict(value = "dailyPayments", allEntries = true)
-    public void deleteDailyPayments(DeleteDailyPaymentsRequestDto request) {
+    public void deleteDailyPayments(
+            String memberEmail, DeleteDailyPaymentsRequestDto request
+    ) {
+        Member member = validateMember(memberEmail);
 
-        Member member = validateMember(request.getMemberEmail());
+        DailyPayments dailyPayments =
+                validateDailyPayments(request.getDailyPaymentsId());
 
-        DailyPayments dailyPayments = validateDailyPayments(request.getDailyPaymentsId());
-
-        requestMemberAndDailyPaymentsOwnerMismatch(member, dailyPayments);
+        checkDailyPaymentsOwner(member, dailyPayments);
 
         dailyPaymentsRepository.delete(dailyPayments);
     }
 
     @Transactional(readOnly = true)
-    public GetDailyPaymentsResponseDto getDailyPayments(Long dailyPaymentsId) {
+    public GetDailyPaymentsResponseDto getDailyPayments(
+            String memberEmail, Long dailyPaymentsId
+    ) {
+
+        Member member = validateMember(memberEmail);
 
         DailyPayments dailyPayments = validateDailyPayments(dailyPaymentsId);
+
+        checkDailyPaymentsOwner(member, dailyPayments);
 
         return GetDailyPaymentsResponseDto.of(dailyPayments);
     }
 
     @Transactional(readOnly = true)
     @Cacheable("dailyPayments")
-    public List<GetDailyPaymentsResponseDto> getDailyPaymentsList() {
+    public List<GetDailyPaymentsResponseDto> getDailyPaymentsList(
+            String memberEmail
+    ) {
+        validateMember(memberEmail);
+
         List<DailyPayments> all = dailyPaymentsRepository.findAll();
 
         return all.stream()
@@ -119,7 +132,7 @@ public class DailyPaymentsService {
                 .collect(Collectors.toList());
     }
 
-    private static void requestMemberAndDailyPaymentsOwnerMismatch(
+    private static void checkDailyPaymentsOwner(
             Member member, DailyPayments dailyPayments
     ) {
         if (!dailyPayments.getMember().getId().equals(member.getId())) {
@@ -153,7 +166,7 @@ public class DailyPaymentsService {
     private Member validateMember(String memberEmail) {
         return memberRepository.findByEmail(memberEmail).orElseThrow(
                 () -> new AccountBookException(
-                        "존재하지 않는 화면입니다.",
+                        "존재하지 않는 회원입니다.",
                         NOT_FOUND_USER_EXCEPTION
                 )
         );
